fix logic

+++ controllers/customers.controller.js
@@ -8,21 +8,21 @@ exports.get = async (req, res) => {
     .populate("favouriteGenres")
     .exec();
 
-  if (!customers.length) return res.status(404).json({ message: "There aren't customers, create a new one!" });
+  if (!customers.length) return res.status(404).json({ error: "There aren't customers, create a new one!" });
 
   res.status(200).json({ customers });
 };
 
 exports.create = async (req, res) => {
   const foundCustomer = await Customer.findOne({ user: req.user.userId });
-  if (foundCustomer) return res.status(400).json({ message: "Customer already exists for this user." });
+  if (foundCustomer) return res.status(400).json({ error: "Customer already exists for this user." });
 
   const { favouriteGenres } = req.body;
   let genreIds = [];
 
   for (let genre of favouriteGenres) {
     const foundGenre = await Genre.findOne({ name: genre });
-    if (!foundGenre) return res.status(400).json({ message: "Invalid genre." });
+    if (!foundGenre) return res.status(400).json({ error: "Invalid genre." });
 
     genreIds.push(foundGenre._id);
   }
@@ -45,14 +45,17 @@ exports.update = async (req, res) => {
 
   for (let genre of favouriteGenres) {
     const foundGenre = await Genre.findOne({ name: genre });
-    if (!foundGenre) return res.status(400).json({ message: "Invalid genre." });
+    if (!foundGenre) return res.status(400).json({ error: "Invalid genre." });
 
     genreIds.push(foundGenre._id);
   }
 
+  const foundCustomer = await Customer.findOne({ "user._id": req.user.userId });
+  if (!foundCustomer) return res.status(403).json({ error: "Unauthorized." });
+
   const customer = await Customer.findByIdAndUpdate(id, { ...req.body, favouriteGenres: genreIds }, { new: true });
 
-  if (!customer) return res.status(404).json({ message: "The customer with the given ID was not found." });
+  if (!customer) return res.status(404).json({ error: "The customer with the given ID was not found." });
 
   res.status(200).json({ customer });
 };
@@ -60,8 +63,11 @@ exports.update = async (req, res) => {
 exports.delete = async (req, res) => {
   const { id } = req.params;
 
+  const foundCustomer = await Customer.findOne({ "user._id": req.user.userId });
+  if (!foundCustomer) return res.status(403).json({ error: "Unauthorized." });
+
   const customer = await Customer.findByIdAndDelete(id);
-  if (!customer) return res.status(404).json({ message: "The customer with the given ID was not found." });
+  if (!customer) return res.status(404).json({ error: "The customer with the given ID was not found." });
 
   res.status(200).json({ customer });
 };
@@ -70,7 +76,7 @@ exports.getSingle = async (req, res) => {
   const { id } = req.params;
   const customer = await Customer.findById(id).populate("favouriteGenres").populate("user");
 
-  if (!customer) return res.status(404).json({ message: "The customer with the given ID was not found." });
+  if (!customer) return res.status(404).json({ error: "The customer with the given ID was not found." });
 
   res.status(200).json({ customer });
 };
